# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChurchOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visitors:
    get:
      operationId: listVisitors
      tags: [visitors]
      parameters:
        - name: church
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Visitor" }
    post:
      operationId: createVisitor
      tags: [visitors]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVisitorBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Visitor" }
  /visitors/{id}/status:
    patch:
      operationId: updateVisitorStatus
      tags: [visitors]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Visitor" }
  /visitors/summary:
    get:
      operationId: getVisitorsSummary
      tags: [visitors]
      parameters:
        - name: church
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VisitorsSummary" }

  # ----- Programs / Bulletins -----
...
  /sermons:
    get:
      operationId: listSermons
      tags: [sermons]
      parameters:
        - name: church
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Sermon" }
    post:
      operationId: createSermon
      tags: [sermons]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSermonBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sermon" }
  /sermons/{id}:
    get:
      operationId: getSermon
      tags: [sermons]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sermon" }
  /sermons/{id}/play:
    post:
      operationId: incrementSermonPlay
      tags: [sermons]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  playCount: { type: integer }
                required: [playCount]

  # ----- Notifications -----
...
components:
  schemas:
...
    Visitor:
      type: object
      required: [id, churchName, fullName, phone, status, firstTime, visitDate, createdAt]
      properties:
        id: { type: integer }
        churchName: { type: string }
        fullName: { type: string }
        phone: { type: string }
        email: { type: string, nullable: true }
        howHeard: { type: string, nullable: true }
        firstTime: { type: boolean }
        visitDate: { type: string }
        status: { type: string }
        cellLeaderEmail: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string }
    CreateVisitorBody:
      type: object
      required: [churchName, fullName, phone]
      properties:
        churchName: { type: string }
        fullName: { type: string }
        phone: { type: string }
        email: { type: string }
        howHeard: { type: string }
        firstTime: { type: boolean }
    VisitorsSummary:
      type: object
      required: [total, newCount, contactedCount, returnedCount, coldCount]
      properties:
        total: { type: integer }
        newCount: { type: integer }
        contactedCount: { type: integer }
        returnedCount: { type: integer }
        coldCount: { type: integer }
...
    Sermon:
      type: object
      required: [id, churchName, title, preacher, sermonDate, scripture, playCount, createdAt]
      properties:
        id: { type: integer }
        churchName: { type: string }
        title: { type: string }
        preacher: { type: string }
        sermonDate: { type: string }
        scripture: { type: string }
        seriesName: { type: string, nullable: true }
        description: { type: string, nullable: true }
        audioFilename: { type: string, nullable: true }
        playCount: { type: integer }
        createdAt: { type: string }
    CreateSermonBody:
      type: object
      required: [churchName, title, preacher, sermonDate, scripture]
      properties:
        churchName: { type: string }
        title: { type: string }
        preacher: { type: string }
        sermonDate: { type: string }
        scripture: { type: string }
        seriesName: { type: string }
        description: { type: string }
        audioFilename: { type: string }
...